Separate TJSGameSettings from its Stores
